Cleaned up Mainwindow and added the readerHelper Class
